Using a custom dispatcher

+++ common/src/main/scala/nl/os3/ls/Messages.scala
@@ -6,9 +6,9 @@ case object Yo extends Messages
 
 sealed trait Tasks
 
-case class IOTasks(taskId:String, count:Int, ioThreadSleep: Int) extends Tasks
+case class IOTasks(taskId:String, count:Int, ioThreadSleep: Long) extends Tasks
 case class CPUTasks(taskId:String, count:Int, fibCompute:Int) extends Tasks
 
-case class ExperimentData(cpuTasks:Int, ioTasks:Int, delay:Int, fibCompute:Int, ioThreadSleep: Int)
+case class ExperimentData(cpuTasks:Int, ioTasks:Int, delay:Int, fibCompute:Int, ioThreadSleep: Long)
 
 case class CreateWorkerRouter(amount:Int)
